chore(CX-40350): enhance logging configuration and introduce test helpers
- Add `elixirc_paths` configuration for different environments in `mix.exs`.
- Introduce a new `PlugTestHelpers` module with various plug implementations for testing, including debug, info, and conditional logging.
- Update tests to utilize the new helper modules for improved organization and clarity.
- Adjust duration assertions in tests to accommodate timing variations.

These changes streamline the logging setup and enhance the testing framework for better maintainability and flexibility.
